requirements.txt
absl_py==2.1.0
accelerate==0.28.0
altair==5.3.0
beautifulsoup4==4.12.3
blinker==1.8.1
blis==0.7.11
cachetools==5.3.3
certifi==2023.7.22
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
contexttimer==0.3.3
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
decord==0.6.0
diffusers==0.25.0
distlib==0.3.8
einops==0.8.0
executing==2.0.1
fairscale==0.4.4
filelock==3.13.1
fonttools==4.51.0
fsspec==2024.3.1
ftfy==6.2.0
gdown==5.1.0
GitPython==3.1.43
grpcio==1.62.2
huggingface_hub==0.22.2
identify==2.5.36
imageio==2.34.1
iopath==0.1.9
ipython==8.24.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
kaggle==1.6.12
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
lightning_utilities==0.11.2
marisa-trie==1.1.0
Markdown==3.6
markdown_it_py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.3
nodeenv==1.8.0
numpy==1.25.2
omegaconf==2.3.0
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
packaging==24.0
pandas==2.2.1
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
Pillow_SIMD==9.0.0.post1
platformdirs==4.2.1
plotly==5.21.0
portalocker==2.8.2
pre-commit==3.7.0
preshed==3.0.9
prompt_toolkit==3.0.43
protobuf==4.25.3
ptyprocess==0.7.0
pure_eval==0.2.2
pyarrow==16.0.0
pycocoevalcap==1.2
pycocotools==2.0.7
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.0
pygments==2.17.2
pyparsing==3.0.9
PySocks==1.7.1
python-magic==0.4.27
python-slugify==8.0.4
python_dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.0
regex==2023.8.8
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.1
salesforce-lavis==1.0.2
scikit-image==0.23.2
scikit_learn==1.3.1
scipy==1.12.0
seaborn==0.13.2
sentence-transformers==2.7.0
sentencepiece==0.2.0
six==1.16.0
smart_open==6.4.0
smmap==5.0.1
soupsieve==2.5
spacy==3.7.2
spacy_legacy==3.0.12
spacy_loggers==1.0.5
srsly==2.4.5
stack_data==0.6.3
streamlit==1.33.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
text-unidecode==1.3
thinc==8.2.1
threadpoolctl==3.5.0
tifffile==2024.4.24
timm==0.4.12
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.2.1
torchmetrics==1.3.2
torchvision==0.17.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
transformers==4.40.1
typer==0.9.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.1.0
virtualenv==20.26.1
wasabi==1.1.2
watchdog==4.0.0
wcwidth==0.2.13
weasel==0.3.4
webdataset==0.2.86
webencodings==0.5.1
Werkzeug==3.0.2
